git Revert "Organizing root/base files to (root) folder."
This reverts commit 08a0e14a9ec72d6994538b53b01080cb69a840db.
